refactor : [#52] 리뷰 반영 네이밍 변경 및 레이어 이동
