+++ src/main/java/ttf/lost/application/avatar/AvatarService.java
@@ -3,8 +3,8 @@ package ttf.lost.application.avatar;
 import java.util.List;
 
 import ttf.lost.infrastructure.api.avatar.AvatarAndPriceDto;
-import ttf.lost.infrastructure.api.avatar.AvatarAndTotalPriceDto;
 import ttf.lost.infrastructure.api.avatar.AvatarDto;
+import ttf.lost.presentation.api.avatar.response.AvatarAndTotalPriceResponse;
 
 public interface AvatarService {
 	/**
@@ -24,5 +24,5 @@ public interface AvatarService {
 	 * 리스트의 각 Price들을 더해 Total을 받고 나머지 정보들과 묶어서 Return
 	 * @param avatarAndPriceDtoList 이전 List에서 받아온 데이터들
 	 */
-	AvatarAndTotalPriceDto avatarTotalPriceAndInfo(List<AvatarAndPriceDto> avatarAndPriceDtoList);
+	AvatarAndTotalPriceResponse avatarTotalPriceAndInfo(List<AvatarAndPriceDto> avatarAndPriceDtoList);
 }
