+++ src/main/java/ttf/lost/application/avatar/AvatarServiceFacade.java
@@ -6,21 +6,18 @@ import org.springframework.stereotype.Service;
 
 import lombok.RequiredArgsConstructor;
 import ttf.lost.infrastructure.api.avatar.AvatarAndPriceDto;
-import ttf.lost.infrastructure.api.avatar.AvatarAndTotalPriceDto;
 import ttf.lost.infrastructure.api.avatar.AvatarDto;
+import ttf.lost.presentation.api.avatar.response.AvatarAndTotalPriceResponse;
 
 @Service
 @RequiredArgsConstructor
 public class AvatarServiceFacade {
 	private final AvatarService avatarService;
 
-	public AvatarAndTotalPriceDto findUserAvatarAndPrice(String nickname) {
-		// TODO : Open API 호출 후 DTO에 저장,
+	public AvatarAndTotalPriceResponse findUserAvatarAndPrice(String nickname) {
 		List<AvatarDto> userAvatarList = avatarService.findUserAvatar(nickname);
-		// TODO : Repository에 정보 및 Open API (거래소 검색) 호출 후 가격 저장 (DTO)
 		List<AvatarAndPriceDto> avatarAndPriceDtoList =
 			avatarService.avatarPriceSave(userAvatarList, nickname);
-		// TODO : 가격을 모두 더해주고 위의 정보들과 같이 Return
 		return avatarService.avatarTotalPriceAndInfo(avatarAndPriceDtoList);
 	}
 }
